fix: RN Intercom on iOS run not on main thread
When running Intercom I noticed that the app is lagging on the first usage of any Intercom method (subsequent
usages seems to be OK). When checking logs in XCode, I noticed that there's warning "- [UIApplication
supportsMultipleScenes] must be used from main thread only" for any method that is run as a first one in the app.

To handle all possible cases, I made a patch where I declare the main queue as a `- methodQueue`
(https://reactnative.dev/docs/native-modules-ios#threading)